task: moved hooks folder to components root dir
